+++ gcd.c
@@ -2,16 +2,16 @@
 
 int main(void)
 {
-    int a,b,tmp;
+    unsigned long long a,b,tmp;
     printf("Greatest Common Divisor Calculator\n");
     printf("Enter two numbers: ");
-    scanf("%d%d",&a,&b);
+    scanf("%u%u",&a,&b);
     while (b != 0)
     {
         tmp = a % b;
         a = b;
         b = tmp;
     }
-    printf("%d\n",a);
+    printf("%u\n",a);
     return 0;
 }
