+++ matrix.c
@@ -3,8 +3,8 @@
 int main(void)
 {
     printf("NumberFlow Program\n");
-    int x;
-    long long int y = 99999999999999;
+    unsigned long long x;
+    unsigned long long int y = 99999999999999;
     for (x = 1;x <= y;x++)
     {
         printf("10100101010010100100101010010100101001001010101001010110101010101010010111010101");
